collect the home ui state in state package
